+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Remove `scheduled` flag from `task list`, the `filter` flag covers this use case now. Use `today & !no time`.
 - Add `filter` flag to `task edit` and `task next`
 - Add `filter label`
+- Add `filter schedule`
+- Add `filter prioritize`
 
 ## 2023-10-09 v0.5.3